requirements.txt
Flask==3.1.1
bcrypt==4.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
Flask-Limiter==3.12
requests==2.32.4
google-genai==1.29.0
google-generativeai==0.8.5
pytz==2025.2
# jwt==1.4.0
PyJWT==2.10.1